snmp_subagent: Call walk_iftable() after walk_nodetable()
When one node has restarted, on the other node, the values of
LHANodeIFCount all become 0.
The reason is because walk_nodetable() function initializes the number
of I/F (LHANodeIFCount) on each node, but these correct value is set in
walk_iftable().
And in NodeStatus() callback, it executed walk_nodetable() only.